cleaned OGL_geometry_data, I need to make a new one to take vertex color into account

+++ build/clean.sh
@@ -1,7 +1,21 @@
 #! /bin/bash
 
-rm -f CMakeCache.txt &&\
-	rm -f MakeFile &&\
-	rm -f -r CMakeFiles &&\
-	cmake CMakeLists.txt
+read -p "Are you sure?  
+This will remove all relevant compilation files ...
+(y or n) " ANSWER
+
+if [ ${ANSWER} == y ];
+then
+
+	rm -f CMakeCache.txt &&\
+		rm -f MakeFile &&\
+		rm -f -r CMakeFiles &&\
+		cmake CMakeLists.txt
+fi
+
+if [ ${ANSWER} == n ];
+then
+	exit 1
+fi
+
